make sure user cannot remove file while uploading to avoid error in gallery or menu in every detail page

+++ resources/views/collaborator/collaborator.blade.php
@@ -3765,6 +3765,7 @@
                         myDropzone.processQueue();
                         $("#btnSaveGallery").html('Uploading Gallery...');
                         $("#btnSaveGallery").addClass('disabled');
+                        $(".btn-del").addClass("btn-del-disable");
                     }
                 });
 
@@ -3805,8 +3806,10 @@
                         e.preventDefault();
                         e.stopPropagation();
 
-                        // Remove the file preview.
-                        _this.removeFile(file);
+                        if (!e.target.classList.contains("btn-del-disable")) {
+                            // Remove the file preview.
+                            _this.removeFile(file);
+                        }
                         // If you want to the delete the file on the server as well,
                         // you can do the AJAX request here.
                     });
